requirements.txt
Flask==3.0.3
torch>=2.1.0
torchvision>=0.16.0
Pillow>=10.0.0
numpy>=1.24.0

